新增回复、删除回复接口.修复了删除接口传入不存在id可以删除成功的BUG

+++ MatsuriBBS/src/main/java/com/cn/matsuribbs/biz/PostBiz.java
@@ -60,8 +60,13 @@ public class PostBiz {
 
     public Result deletePost(Integer id) {
         try {
-            postMapper.deleteFun(id);
-            return ResultFactory.buildSuccessResult("删除成功");
+            Post post = postMapper.selectByPostId(id);
+            if (post != null){
+                postMapper.deleteFun(id);
+                return ResultFactory.buildSuccessResult("删除成功");
+            } else {
+                return ResultFactory.buildFailResult("删除失败,未找到相应帖子");
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return ResultFactory.buildFailResult("删除失败");
